Logger, Thread windows export fix.
Logger and Thread are exported under windows in a more proper manner.
