+++ src/Logger.cpp
@@ -1,6 +1,3 @@
-#include <cassert>
-
-#include <stdarg.h>
 #if defined(WIN32)
 #else
 #include <cxxabi.h>
@@ -8,11 +5,9 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include <iostream>
-#include <vector>
 #include <regex>
 
 #include "Logger.h"
-#include "Assert.h"
 #include "TraceListener.h"
 #include "Process.h"
 #include "Environment.h"
@@ -85,10 +80,10 @@ namespace core
         m_loggerImpl->Start(severity);
         m_severity = severity;
     }
-    void Logger::Log(TraceSeverity severity, const string& message)
+    void Logger::Log(TraceSeverity severity, const char* message)
     {
         assert(m_loggerImpl.get() != nullptr);
-        m_loggerImpl->Log(severity, message.c_str());
+        m_loggerImpl->Log(severity, message);
     }
 
     void Logger::Flush()
@@ -108,54 +103,4 @@ namespace core
         m_loggerImpl.release();
     }
 
-    string Logger::BuildMessage(const Source& source, const char* format, ...)
-    {
-        va_list arguments;
-        va_start(arguments, format);
-        string result;
-
-        char buf[Local_buffer_size] = "";
-#if defined(WIN32)
-        int size = _snprintf(buf, Local_buffer_size, "%s:%s:%d\t", source.file,
-            source.function, source.line);
-#else
-        int size;
-        PLATFORM_VERIFY((size = snprintf(buf, Local_buffer_size, "%s:%s:%d\t", source.file, source.function, source.line)) >= 0);
-#endif
-        assert(size != -1 && size < Local_buffer_size); //In windows version -1 is a legit answer
-#if defined(WIN32)
-        int tempSize = vsnprintf(buf + size, Local_buffer_size - size, format, arguments);
-        tempSize != -1 ? size += tempSize : size = -1;
-#else
-        int tempSize;
-        PLATFORM_VERIFY(tempSize = vsnprintf(buf + size, Local_buffer_size - size, format, arguments) >= 0);
-        size += tempSize;
-#endif
-        
-        if(size != -1 && size < Local_buffer_size)
-            result = buf;
-        else //message was trunced or operation failed
-        {
-            int bufferSize = max(size, 32 * 1024);
-            vector<char> largerBuf;
-            largerBuf.resize(bufferSize);
-            
-#if defined(WIN32)
-            int largerSize = _snprintf(&largerBuf[0], bufferSize, "%s:%s:%d\t", source.file, source.function, source.line);
-#else
-            int largerSize;
-            PLATFORM_VERIFY(largerSize = snprintf(&largerBuf[0], bufferSize, "%s:%s:%d\t", source.file, source.function, source.line) >= 0);
-#endif
-            assert(largerSize != -1 && largerSize < bufferSize); //In windows version -1 is a legit answer
-#if defined(WIN32)
-            vsnprintf(&largerBuf[largerSize], bufferSize - largerSize, format, arguments); //We will print what we can, no second resize.
-#else
-            PLATFORM_VERIFY(vsnprintf(&largerBuf[largerSize], bufferSize - largerSize, format, arguments) >= 0); //We will print what we can, no second resize.
-#endif
-            result = string(largerBuf.begin(), largerBuf.end());
-        }
-
-        va_end(arguments);
-        return result;
-    }
 }
