--- src/Thread.cpp
@@ -1,40 +1,0 @@
-#include <cassert>
-#include <iostream>
-#include "Thread.h"
-#include "Assert.h"
-
-using namespace std;
-
-namespace core
-{
-    Thread::Thread(const string& threadName, const function<void(void)>& requestedPoint)
-        : m_name(threadName), m_requestedPoint(requestedPoint)
-    {    
-        assert((int)m_name.size() <= MAX_THREAD_NAME);
-    }
-
-    void Thread::Start()
-    {
-        m_thread.reset(new thread(bind(&Thread::EntryPoint, this)));
-    }
-
-    void Thread::Join() const
-    {
-        m_thread->join();
-    }
-
-    void Thread::EntryPoint()
-    {
-        try
-        {
-            m_requestedPoint();
-        }
-        catch(const Exception&)
-        {
-        }
-        catch(std::exception& e)
-        {
-            TRACE_ERROR("%s", e.what());
-        }
-    }
-}
